+++ projects/exui-common-lib/src/lib/components/selected-case-list/selected-case-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { SharedCase } from '../../models/case-share.module';
+import { SharedCase } from '../../models/case-share.model';
 
 @Component({
   selector: 'xuilib-selected-case-list',
