+++ projects/exui-common-lib/src/lib/models/case-share.model.ts
@@ -1,4 +1,4 @@
-import { UserDetails } from './user-details.module';
+import { UserDetails } from './user-details.model';
 
 export interface SharedCase {
   caseId: string;
