+++ projects/exui-common-lib/src/lib/models/user-details.model.ts
